Update README.md
Code Obfuscation being an upcoming field attracts, and provides us with a new different way of exercising security and making it more difficult to reverse engineer the source code and also difficult to maintain and modify to extents desired.
The project aimed at designing a software which obfuscates the source codes written in C programming language, while the software is written in Java programming language.
The process that involves the construction of the parse forms the most critical step of the project, is also displayed in the software. This basically provides us with a standard form of the given source code, and techniques can be well implemented on the parse tree, traversing through it, which again involves coming up with newer standards.
The system constructed in this project performs the lexical analysis of the input source code and constructs its parse tree, with the constraints mentioned. The obfuscation is achieved at the level of variable replacement in the main() function, which is known as Layout Obfuscation. The proposed project could not be completed due to extensive study carried out in the designing of the lexical analysis module and the parse tree construction module, and also due to time constraint. It is believed that this system if extended can prove to be a useful Code Obfuscator.